requirements.txt
dash==2.17.1
dash_cytoscape==1.0.1
langchain==0.2.5
langchain-community==0.2.5
matplotlib==3.8.3
neo4j==5.21.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
tensorboard==2.17.0
tokenizers==0.19.1
torch==2.3.1
torchvision==2.3.1
torchaudio==2.3.1
torchviz==0.0.2
tqdm==4.66.2
transformers==4.41.2
wikipedia==1.4.0
python-dotenv